+++ DIBBLE-YAMILA_VIQUE-FABRIZIO/VIQUE-FABRIZIO_DIBBLE-YAMILA/src/main/java/com/backend/clinica/dto/salida/PacienteSalidaDto.java
@@ -2,7 +2,7 @@ package com.backend.clinica.dto.salida;
 import java.time.LocalDate;
 
 public class PacienteSalidaDto {
-    private int id;
+    private Long id;
     private String nombre;
     private String apellido;
     private int dni;
@@ -12,7 +12,7 @@ public class PacienteSalidaDto {
     public PacienteSalidaDto() {
     }
 
-    public PacienteSalidaDto(int id, String nombre, String apellido, int dni, LocalDate fechaIngreso, DomicilioSalidaDto domicilioSalidaDto) {
+    public PacienteSalidaDto(Long id, String nombre, String apellido, int dni, LocalDate fechaIngreso, DomicilioSalidaDto domicilioSalidaDto) {
         this.id = id;
         this.nombre = nombre;
         this.apellido = apellido;
@@ -21,11 +21,11 @@ public class PacienteSalidaDto {
         this.domicilioSalidaDto = domicilioSalidaDto;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
